refactor: split topicStore
