+++ tests/test-basic-topic-store.test.ts
@@ -1,5 +1,7 @@
 import { describe, it, expect, beforeAll, beforeEach } from 'vitest';
 import { useTopicStore } from '../apps/client/src/store/topicStore';
+import { useGoalStore } from '../apps/client/src/store/goalStore';
+import { useTaskStore } from '../apps/client/src/store/taskStore';
 import { initTestAuth } from '../vitest.setup';
 
 describe('Topic Store - Basic Tests', () => {
@@ -16,8 +18,7 @@ describe('Topic Store - Basic Tests', () => {
 
   beforeEach(() => {
     // 確保每個測試開始時都有乾淨的狀態
-    const store = useTopicStore.getState();
-    store.reset();
+    useTopicStore.getState().reset();
   });
 
   it('應該有正確的初始狀態', () => {
@@ -31,64 +32,67 @@ describe('Topic Store - Basic Tests', () => {
   });
 
   it('應該有所有必要的方法', () => {
-    const store = useTopicStore.getState();
-    
+    const topic = useTopicStore.getState();
+    const goal = useGoalStore.getState();
+    const task = useTaskStore.getState();
+
     // 核心 CRUD 方法
-    expect(typeof store.fetchTopics).toBe('function');
-    expect(typeof store.getTopic).toBe('function');
-    expect(typeof store.createTopic).toBe('function');
-    expect(typeof store.updateTopic).toBe('function');
-    expect(typeof store.deleteTopic).toBe('function');
-    
+    expect(typeof topic.fetchTopics).toBe('function');
+    expect(typeof topic.getTopic).toBe('function');
+    expect(typeof topic.createTopic).toBe('function');
+    expect(typeof topic.updateTopic).toBe('function');
+    expect(typeof topic.deleteTopic).toBe('function');
+
     // Goals 操作
-    expect(typeof store.addGoal).toBe('function');
-    expect(typeof store.updateGoal).toBe('function');
-    expect(typeof store.deleteGoal).toBe('function');
-    
+    expect(typeof goal.addGoal).toBe('function');
+    expect(typeof goal.updateGoal).toBe('function');
+    expect(typeof goal.deleteGoal).toBe('function');
+
     // Tasks 操作
-    expect(typeof store.addTask).toBe('function');
-    expect(typeof store.updateTask).toBe('function');
-    expect(typeof store.deleteTask).toBe('function');
-    
+    expect(typeof task.addTask).toBe('function');
+    expect(typeof task.updateTask).toBe('function');
+    expect(typeof task.deleteTask).toBe('function');
+
     // 還原功能
-    expect(typeof store.restoreTopic).toBe('function');
-    expect(typeof store.restoreGoal).toBe('function');
-    expect(typeof store.restoreTask).toBe('function');
-    
+    expect(typeof topic.restoreTopic).toBe('function');
+    expect(typeof goal.restoreGoal).toBe('function');
+    expect(typeof task.restoreTask).toBe('function');
+
     // 兼容性方法
-    expect(typeof store.updateTopicCompat).toBe('function');
-    expect(typeof store.updateGoalCompat).toBe('function');
-    expect(typeof store.updateTaskCompat).toBe('function');
-    
+    expect(typeof topic.updateTopicCompat).toBe('function');
+    expect(typeof goal.updateGoalCompat).toBe('function');
+    expect(typeof task.updateTaskCompat).toBe('function');
+
     // 協作功能
-    expect(typeof store.enableTopicCollaboration).toBe('function');
-    expect(typeof store.disableTopicCollaboration).toBe('function');
+    expect(typeof topic.enableTopicCollaboration).toBe('function');
+    expect(typeof topic.disableTopicCollaboration).toBe('function');
   });
 
   it('應該有版本控制相關的方法', () => {
-    const store = useTopicStore.getState();
-    
-    expect(typeof store.markTaskCompleted).toBe('function');
-    expect(typeof store.markTaskInProgress).toBe('function');
-    expect(typeof store.markTaskTodo).toBe('function');
+    const task = useTaskStore.getState();
+
+    expect(typeof task.markTaskCompleted).toBe('function');
+    expect(typeof task.markTaskInProgress).toBe('function');
+    expect(typeof task.markTaskTodo).toBe('function');
   });
 
   it('應該有快速查詢方法', () => {
-    const store = useTopicStore.getState();
-    
-    expect(typeof store.getActiveTasksForUser).toBe('function');
-    expect(typeof store.getTopicWithStructure).toBe('function');
+    const topic = useTopicStore.getState();
+    const task = useTaskStore.getState();
+
+    expect(typeof task.getActiveTasksForUser).toBe('function');
+    expect(typeof topic.getTopicWithStructure).toBe('function');
   });
 
   it('應該有任務動作相關的方法', () => {
-    const store = useTopicStore.getState();
-    
-    expect(typeof store.performTaskAction).toBe('function');
-    expect(typeof store.checkInTask).toBe('function');
-    expect(typeof store.addTaskCount).toBe('function');
-    expect(typeof store.addTaskAmount).toBe('function');
-    expect(typeof store.resetTaskProgress).toBe('function');
-    expect(typeof store.cancelTodayCheckIn).toBe('function');
+    const task = useTaskStore.getState();
+
+    expect(typeof task.performTaskAction).toBe('function');
+    expect(typeof task.checkInTask).toBe('function');
+    expect(typeof task.addTaskCount).toBe('function');
+    expect(typeof task.addTaskAmount).toBe('function');
+    expect(typeof task.resetTaskProgress).toBe('function');
+    expect(typeof task.cancelTodayCheckIn).toBe('function');
   });
 
   it('應該能正確設置和重置狀態', () => {
