+++ tests/test-topic-store.test.ts
@@ -1,12 +1,16 @@
 import { describe, it, expect, beforeEach, afterEach, beforeAll, afterAll } from 'vitest';
 import { useTopicStore } from '../apps/client/src/store/topicStore';
+import { useGoalStore } from '../apps/client/src/store/goalStore';
+import { useTaskStore } from '../apps/client/src/store/taskStore';
 import { SUBJECTS } from '../apps/client/src/constants/subjects';
 import { Topic } from '../apps/client/src/types/goal';
 import { initTestAuth, cleanupTestData } from '../vitest.setup';
 import { supabase } from '../apps/client/src/services/supabase';
 
 describe('TopicStore', () => {
-  let store: ReturnType<typeof useTopicStore.getState>;
+  let topicStore: ReturnType<typeof useTopicStore.getState>;
+  let goalStore: ReturnType<typeof useGoalStore.getState>;
+  let taskStore: ReturnType<typeof useTaskStore.getState>;
   let createdTopics: string[] = [];
   
   beforeAll(async () => {
@@ -16,15 +20,17 @@ describe('TopicStore', () => {
   });
 
   beforeEach(() => {
-    store = useTopicStore.getState();
-    store.reset();
+    topicStore = useTopicStore.getState();
+    goalStore = useGoalStore.getState();
+    taskStore = useTaskStore.getState();
+    topicStore.reset();
   });
 
   afterEach(async () => {
     // 清理測試資料
     for (const id of createdTopics) {
       try {
-        await store.deleteTopic(id);
+        await topicStore.deleteTopic(id);
       } catch (error) {
         console.warn(`清理主題 ${id} 失敗:`, error);
       }
@@ -52,7 +58,7 @@ describe('TopicStore', () => {
     };
 
     it('應該能創建新主題', async () => {
-      const result = await store.createTopic(mockTopic);
+      const result = await topicStore.createTopic(mockTopic);
       expect(result).toBeDefined();
       expect(result?.id).toBeDefined();
       expect(result?.title).toBe(mockTopic.title);
@@ -61,7 +67,7 @@ describe('TopicStore', () => {
       if (result?.id) createdTopics.push(result.id);
 
       // 驗證資料庫中的資料
-      const savedTopic = await store.getTopic(result!.id);
+      const savedTopic = await topicStore.getTopic(result!.id);
       expect(savedTopic).toBeDefined();
       expect(savedTopic?.title).toBe(mockTopic.title);
       expect(savedTopic?.goals).toEqual([]);
@@ -69,33 +75,33 @@ describe('TopicStore', () => {
     });
 
     it('應該能更新主題', async () => {
-      const topic = await store.createTopic(mockTopic);
+      const topic = await topicStore.createTopic(mockTopic);
       if (topic?.id) createdTopics.push(topic.id);
       expect(topic).toBeDefined();
       
       const updatedTitle = '更新後的主題';
-      const result = await store.updateTopicCompat(topic!.id, { title: updatedTitle });
+      const result = await topicStore.updateTopicCompat(topic!.id, { title: updatedTitle });
       expect(result).toBeDefined();
       expect(result?.title).toBe(updatedTitle);
       expect(result?.goals).toEqual([]);
       expect(result?.bubbles).toEqual([]);
 
       // 驗證資料庫中的資料
-      const savedTopic = await store.getTopic(topic!.id);
+      const savedTopic = await topicStore.getTopic(topic!.id);
       expect(savedTopic).toBeDefined();
       expect(savedTopic?.title).toBe(updatedTitle);
     });
 
     it('應該能刪除主題（歸檔）', async () => {
-      const topic = await store.createTopic(mockTopic);
+      const topic = await topicStore.createTopic(mockTopic);
       if (topic?.id) createdTopics.push(topic.id);
       expect(topic).toBeDefined();
       
-      const result = await store.deleteTopic(topic!.id);
+      const result = await topicStore.deleteTopic(topic!.id);
       expect(result).toBe(true);
       
       // 主題應該被歸檔而不是完全刪除
-      const archivedTopic = await store.getTopic(topic!.id);
+      const archivedTopic = await topicStore.getTopic(topic!.id);
       expect(archivedTopic).toBeNull(); // 因為查詢會過濾掉 archived 狀態的記錄
       
       // 驗證資料庫中主題仍存在但狀態為 archived
@@ -109,19 +115,19 @@ describe('TopicStore', () => {
     });
 
     it('應該能還原歸檔的主題', async () => {
-      const topic = await store.createTopic(mockTopic);
+      const topic = await topicStore.createTopic(mockTopic);
       if (topic?.id) createdTopics.push(topic.id);
       expect(topic).toBeDefined();
       
       // 先歸檔主題
-      await store.deleteTopic(topic!.id);
+      await topicStore.deleteTopic(topic!.id);
       
       // 然後還原主題
-      const restoreResult = await store.restoreTopic(topic!.id);
+      const restoreResult = await topicStore.restoreTopic(topic!.id);
       expect(restoreResult).toBe(true);
       
       // 驗證主題已還原
-      const restoredTopic = await store.getTopic(topic!.id);
+      const restoredTopic = await topicStore.getTopic(topic!.id);
       expect(restoredTopic).toBeDefined();
       expect(restoredTopic?.status).toBe('active');
     });
@@ -131,7 +137,7 @@ describe('TopicStore', () => {
     let testTopic;
 
     beforeEach(async () => {
-      testTopic = await store.createTopic({
+      testTopic = await topicStore.createTopic({
         title: '測試主題',
         description: '這是一個測試主題',
         subject: SUBJECTS.MATH,
@@ -158,13 +164,13 @@ describe('TopicStore', () => {
         tasks: []
       };
 
-      const result = await store.addGoal(testTopic!.id, goal);
+      const result = await goalStore.addGoal(testTopic!.id, goal);
       expect(result).toBeDefined();
       expect(result?.title).toBe(goal.title);
       expect(result?.tasks).toEqual([]);
 
       // 驗證資料庫中的資料
-      const savedTopic = await store.getTopic(testTopic!.id);
+      const savedTopic = await topicStore.getTopic(testTopic!.id);
       expect(savedTopic).toBeDefined();
       expect(savedTopic?.goals).toHaveLength(1);
       expect(savedTopic?.goals?.[0].title).toBe(goal.title);
@@ -172,7 +178,7 @@ describe('TopicStore', () => {
     });
 
     it('應該能更新目標', async () => {
-      const goal = await store.addGoal(testTopic!.id, {
+      const goal = await goalStore.addGoal(testTopic!.id, {
         title: '測試目標',
         description: '這是一個測試目標',
         status: 'todo' as const,
@@ -183,20 +189,20 @@ describe('TopicStore', () => {
       expect(goal).toBeDefined();
 
       const updatedTitle = '更新後的目標';
-      const result = await store.updateGoalCompat(testTopic!.id, goal!.id, { title: updatedTitle });
+      const result = await goalStore.updateGoalCompat(testTopic!.id, goal!.id, { title: updatedTitle });
       expect(result).toBeDefined();
       expect(result?.title).toBe(updatedTitle);
       expect(result?.tasks).toEqual([]);
 
       // 驗證資料庫中的資料
-      const savedTopic = await store.getTopic(testTopic!.id);
+      const savedTopic = await topicStore.getTopic(testTopic!.id);
       expect(savedTopic).toBeDefined();
       expect(savedTopic?.goals).toHaveLength(1);
       expect(savedTopic?.goals?.[0].title).toBe(updatedTitle);
     });
 
     it('應該能刪除目標（歸檔）', async () => {
-      const goal = await store.addGoal(testTopic!.id, {
+      const goal = await goalStore.addGoal(testTopic!.id, {
         title: '測試目標',
         description: '這是一個測試目標',
         status: 'todo' as const,
@@ -206,11 +212,11 @@ describe('TopicStore', () => {
       });
       expect(goal).toBeDefined();
 
-      const result = await store.deleteGoal(goal!.id);
+      const result = await goalStore.deleteGoal(goal!.id);
       expect(result).toBe(true);
 
       // 驗證目標被歸檔而不是完全刪除
-      const savedTopic = await store.getTopic(testTopic!.id);
+      const savedTopic = await topicStore.getTopic(testTopic!.id);
       expect(savedTopic).toBeDefined();
       expect(savedTopic?.goals).toHaveLength(0); // 查詢會過濾掉 archived 狀態的記錄
       
@@ -225,7 +231,7 @@ describe('TopicStore', () => {
     });
 
     it('應該能還原歸檔的目標', async () => {
-      const goal = await store.addGoal(testTopic!.id, {
+      const goal = await goalStore.addGoal(testTopic!.id, {
         title: '測試目標',
         description: '這是一個測試目標',
         status: 'todo' as const,
@@ -236,14 +242,14 @@ describe('TopicStore', () => {
       expect(goal).toBeDefined();
 
       // 先歸檔目標
-      await store.deleteGoal(goal!.id);
+      await goalStore.deleteGoal(goal!.id);
 
       // 然後還原目標
-      const restoreResult = await store.restoreGoal(goal!.id);
+      const restoreResult = await goalStore.restoreGoal(goal!.id);
       expect(restoreResult).toBe(true);
 
       // 驗證目標已還原
-      const savedTopic = await store.getTopic(testTopic!.id);
+      const savedTopic = await topicStore.getTopic(testTopic!.id);
       expect(savedTopic).toBeDefined();
       expect(savedTopic?.goals).toHaveLength(1);
       expect(savedTopic?.goals?.[0].status).toBe('todo');
@@ -255,7 +261,7 @@ describe('TopicStore', () => {
     let testGoal;
 
     beforeEach(async () => {
-      testTopic = await store.createTopic({
+      testTopic = await topicStore.createTopic({
         title: '測試主題',
         description: '這是一個測試主題',
         subject: SUBJECTS.MATH,
@@ -270,7 +276,7 @@ describe('TopicStore', () => {
       if (testTopic?.id) createdTopics.push(testTopic.id);
       expect(testTopic).toBeDefined();
 
-      testGoal = await store.addGoal(testTopic!.id, {
+      testGoal = await goalStore.addGoal(testTopic!.id, {
         title: '測試目標',
         description: '這是一個測試目標',
         status: 'todo' as const,
@@ -292,19 +298,19 @@ describe('TopicStore', () => {
         dueDate: new Date().toISOString()
       };
 
-      const result = await store.addTask(testGoal!.id, task);
+      const result = await taskStore.addTask(testGoal!.id, task);
       expect(result).toBeDefined();
       expect(result?.title).toBe(task.title);
 
       // 驗證資料庫中的資料
-      const savedTopic = await store.getTopic(testTopic!.id);
+      const savedTopic = await topicStore.getTopic(testTopic!.id);
       expect(savedTopic).toBeDefined();
       expect(savedTopic?.goals?.[0].tasks).toHaveLength(1);
       expect(savedTopic?.goals?.[0].tasks?.[0].title).toBe(task.title);
     });
 
     it('應該能更新任務狀態', async () => {
-      const task = await store.addTask(testGoal!.id, {
+      const task = await taskStore.addTask(testGoal!.id, {
         title: '測試任務',
         description: '這是一個測試任務',
         status: 'todo' as const,
@@ -315,18 +321,18 @@ describe('TopicStore', () => {
       });
       expect(task).toBeDefined();
 
-      const result = await store.updateTaskCompat(testTopic!.id, testGoal!.id, task!.id, { status: 'done' });
+      const result = await taskStore.updateTaskCompat(testTopic!.id, testGoal!.id, task!.id, { status: 'done' });
       expect(result).toBeDefined();
       expect(result?.status).toBe('done');
 
       // 驗證資料庫中的資料
-      const savedTopic = await store.getTopic(testTopic!.id);
+      const savedTopic = await topicStore.getTopic(testTopic!.id);
       expect(savedTopic).toBeDefined();
       expect(savedTopic?.goals?.[0].tasks?.[0].status).toBe('done');
     });
 
     it('應該能刪除任務（歸檔）', async () => {
-      const task = await store.addTask(testGoal!.id, {
+      const task = await taskStore.addTask(testGoal!.id, {
         title: '測試任務',
         description: '這是一個測試任務',
         status: 'todo' as const,
@@ -337,11 +343,11 @@ describe('TopicStore', () => {
       });
       expect(task).toBeDefined();
 
-      const result = await store.deleteTask(task!.id);
+      const result = await taskStore.deleteTask(task!.id);
       expect(result).toBe(true);
 
       // 驗證任務被歸檔而不是完全刪除
-      const savedTopic = await store.getTopic(testTopic!.id);
+      const savedTopic = await topicStore.getTopic(testTopic!.id);
       expect(savedTopic).toBeDefined();
       expect(savedTopic?.goals?.[0].tasks).toHaveLength(0); // 查詢會過濾掉 archived 狀態的記錄
       
@@ -356,7 +362,7 @@ describe('TopicStore', () => {
     });
 
     it('應該能還原歸檔的任務', async () => {
-      const task = await store.addTask(testGoal!.id, {
+      const task = await taskStore.addTask(testGoal!.id, {
         title: '測試任務',
         description: '這是一個測試任務',
         status: 'todo' as const,
@@ -368,14 +374,14 @@ describe('TopicStore', () => {
       expect(task).toBeDefined();
 
       // 先歸檔任務
-      await store.deleteTask(task!.id);
+      await taskStore.deleteTask(task!.id);
 
       // 然後還原任務
-      const restoreResult = await store.restoreTask(task!.id);
+      const restoreResult = await taskStore.restoreTask(task!.id);
       expect(restoreResult).toBe(true);
 
       // 驗證任務已還原
-      const savedTopic = await store.getTopic(testTopic!.id);
+      const savedTopic = await topicStore.getTopic(testTopic!.id);
       expect(savedTopic).toBeDefined();
       expect(savedTopic?.goals?.[0].tasks).toHaveLength(1);
       expect(savedTopic?.goals?.[0].tasks?.[0].status).toBe('todo');
@@ -386,7 +392,7 @@ describe('TopicStore', () => {
     let testTopic: Topic | null = null;
 
     beforeEach(async () => {
-      testTopic = await store.createTopic({
+      testTopic = await topicStore.createTopic({
         title: '測試主題',
         description: '這是一個測試主題',
         subject: SUBJECTS.MATH,
@@ -403,23 +409,23 @@ describe('TopicStore', () => {
     });
 
     it('應該能啟用協作狀態', async () => {
-      const result = await store.enableTopicCollaboration(testTopic!.id);
+      const result = await topicStore.enableTopicCollaboration(testTopic!.id);
       expect(result).toBeDefined();
 
-      const savedTopic = await store.getTopic(testTopic!.id);
+      const savedTopic = await topicStore.getTopic(testTopic!.id);
       expect(savedTopic).toBeDefined();
       expect(savedTopic?.is_collaborative).toBe(true);
     });
 
     it('應該能停用協作狀態', async () => {
       // 先啟用協作
-      await store.enableTopicCollaboration(testTopic!.id);
+      await topicStore.enableTopicCollaboration(testTopic!.id);
       
       // 再停用協作
-      const result = await store.disableTopicCollaboration(testTopic!.id);
+      const result = await topicStore.disableTopicCollaboration(testTopic!.id);
       expect(result).toBeDefined();
 
-      const savedTopic = await store.getTopic(testTopic!.id);
+      const savedTopic = await topicStore.getTopic(testTopic!.id);
       expect(savedTopic).toBeDefined();
       expect(savedTopic?.is_collaborative).toBe(false);
     });
